Include the data file, with permission
Includes the data file in the distribution, to make sure that it is
available if the download location changes, or the data service is
temporarily unavailable or ends for some reason.

Adds the "license" note to the readme and includes a file format
description to have it available whenever anyone may need it.