Re-write update mechanism and tests
Uses the smaller 5-year data file for frequent updates and provides a
(hopefully) easy-to-use update function.

Exports the age and update function as well as the data file path in the
hope that they may be useful also outside the package.

Updates the tests to test with and without update, probably overkill,
but you never know.