tests: Separate age test for all and 5-year data
The "all" data file gets updated less often, so the test to be not older
than 100d may fail for that particular file. Currently that age is
hard-coded to 4 years (1460d) to reduce download traffic.
Updates the test to check both files separately, the 5-year data file
for < 100d, and the "all" data file for < 1460d.