ci: Pre-install `pandas`
Pandas might depend on a different `numpy` version than our package.
This indirect dependence is not captured when installing with `-e .`,
which breaks some builds because `pandas` complains about `numpy` being
too old when running the tests.

Pre-installs pandas to resolve that conflict before the package install,
instead of capping the `pandas` version in the package's dependences.