tests: Remove file age testing in gfz tests
This depends too much on the local circumstances and we do not download
or update anything while testing. So the last entry of the file might
lag well behind the current date and 10 days is not much.
Removes this assertion for now until we find a better way to test that.