docs: Include linked license and format files
Includes the top dir in the docs which contains the data licenses and
format descriptions.
There are a bunch of references to local files in the README, and when
that is included in the docs, these links won't be resolved otherwise.
Tries to exclude as much as possible, but still this approach copies the
readme markdown source and the index rst source too.

Removes the static path from the config, which is currently not used.

[skip ci]