gfz: Make daily data reading more flexible
Adds some flexibility to the GFZ daily data reading by passing the
parsing function as an additional argument. The main purpose here is to
(pandas) concatenate the full time series and the nowcast, which is
independent of the actual structure of the `pandas.DataFrame`.
Passing the parsing function enables us to swap out `read_gfz()` in
favour of `read_gfz_wdc()`, for example, or other (future) parsers.