gfz: Replace function argument by dictionary
Replaces passing the parsing function by a dictionary mapping file
formats to parse and update functions. For Hp30 and Hp60, not only the
file names are different, but also the URLS. The URLS are handled
automatically by the Hpo update functions, and we can avoid passing too
many variables to a supposedly simple to use function.